+++ CHANGELOG.md
@@ -6,6 +6,10 @@ All notable changes to the Imperative package will be documented in this file.
 
 - BugFix: Reduced load time by searching for command definitions with `fast-glob` instead of `glob`.
 
+## `5.13.1`
+
+- BugFix: Removed validation of the deprecated pluginHealthCheck property. [#980](https://github.com/zowe/imperative/issues/980)
+
 ## `5.13.0`
 
 - Enhancement: Alters TextUtils behavior slightly to enable daemon color support without TTY